change step progress bar to show green when the goal is reached.
Fix bug regarding sending step count. Before when step data was sent the initial step count was send with the values from the senders device. Which made the receiving devices interpreter the step count falsely. The initial step count is the saved steps in the phone which is different for each device. Now the initialStepCount is sent as -1 which is the default value for the app. Meaning the app fetches the initial step count from the device it self and avoiding the error.